requirements.txt
# Core dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0

# File handling
python-magic-bin==0.4.14
Pillow==10.0.0

# OCR and text processing
pytesseract==0.3.10
pdf2image==1.16.3
python-docx==0.8.11
PyPDF2==3.0.1

# Audio/Video processing
moviepy==1.0.3
SpeechRecognition==3.10.0
pydub==0.25.1

# NLP and ML
transformers==4.31.0
torch==2.0.1
scikit-learn==1.3.0
spacy==3.6.1
nltk==3.8.1
textstat==0.7.3
networkx==3.1

# Visualization
pyvis==0.3.2
plotly==5.16.1

# Database
SQLAlchemy==2.0.20
alembic==1.11.3

# Security
PyJWT==2.8.0
bcrypt==4.0.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
